Added scene class and class creation on GPU

+++ ray.h
@@ -12,7 +12,7 @@ public:
     __device__ point3 origin() const { return orig; }
     __device__ vec3 direction() const { return dir; }
 
-    __device__ point3 at(double t) const {
+    __device__ point3 at(float t) const {
         return orig + t * dir;
     }
